requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
pymongo==4.6.2
dnspython==2.6.1
python-dotenv==1.0.1
pyotp==2.9.0
clamd==1.0.2
waitress==3.0.0 